requirements.txt
deap>=1.3.1
numpy>=1.19.0
smartsheet-python-sdk>=2.101.0
matplotlib>=3.3.0
six>=1.15.0
python-dotenv>=0.15.0
